// src/membuffers/types.d.ts
export type FieldType = number;

export const FieldTypes: {
  TypeMessage: FieldType;
  TypeBytes: FieldType;
  TypeString: FieldType;
  TypeUnion: FieldType;
  TypeUint8: FieldType;
  TypeUint16: FieldType;
  TypeUint32: FieldType;
  TypeUint64: FieldType;
  TypeUint8Array: FieldType;
  TypeUint16Array: FieldType;
  TypeUint32Array: FieldType;
  TypeUint64Array: FieldType;
  TypeMessageArray: FieldType;
  TypeBytesArray: FieldType;
  TypeStringArray: FieldType;
};